tests/test_framework_validation.py
"""
Simple validation tests to check test framework setup.
These tests don't require the full application dependencies.
"""
import pytest


class TestFrameworkValidation:
    """Basic tests to validate the testing framework."""
    
    def test_pytest_working(self):
        """Test that pytest is working correctly."""
        assert True
        
    def test_basic_math(self):
        """Test basic assertions."""
        assert 2 + 2 == 4
        assert "hello" in "hello world"
        
    def test_mock_working(self):
        """Test that mock functionality is available."""
        from unittest.mock import Mock, patch
        
        mock_obj = Mock()
        mock_obj.method.return_value = "test"
        assert mock_obj.method() == "test"
        
    def test_flask_available(self):
        """Test that Flask is available for testing."""
        try:
            import flask
            assert hasattr(flask, 'Flask')
        except ImportError:
            pytest.skip("Flask not available")
            
    def test_environment_isolation(self):
        """Test that tests run in isolated environment."""
        import os
        # This should not conflict with any real environment
        test_var = os.environ.get('TEST_ISOLATION_CHECK', 'not_set')
        assert test_var == 'not_set'  # Should be clean environment


class TestEndpointStructure:
    """Tests to validate the endpoint testing structure."""
    
    def test_route_list_completeness(self):
        """Test that we have identified all the main routes."""
        # Based on our grep search, these are the main endpoints
        expected_routes = [
            # Index routes
            ('/', 'GET'),
            ('/welcome', 'GET'),
            ('/lb-radio', 'GET'),
            ('/lb-radio', 'POST'),
            ('/playlist/create', 'POST'),
            ('/weekly-jams', 'GET'),
            ('/weekly-jams', 'POST'),
            ('/top-tags', 'GET'),
            ('/tag/<tag>', 'GET'),
            ('/unresolved', 'GET'),
            
            # Service routes (prefixed with /service)
            ('/service/', 'GET'),
            ('/service/list', 'GET'),
            ('/service/add', 'GET'),
            ('/service/add', 'POST'),
            ('/service/<slug>/edit', 'GET'),
            ('/service/<slug>/delete', 'GET'),
            ('/service/<slug>/sync', 'GET'),
            ('/service/<slug>/sync/start', 'POST'),
            ('/service/<slug>/sync/start/metadata-only', 'POST'),
            ('/service/<slug>/sync/log', 'GET'),
            ('/service/<slug>/sync/full-log', 'GET'),
            
            # Credential routes (prefixed with /credential)
            ('/credential/list', 'GET'),
            ('/credential/add', 'GET'),
            ('/credential/add', 'POST'),
            ('/credential/<id>/edit', 'GET'),
            ('/credential/<id>/delete', 'GET'),
            
            # Auth routes
            ('/login', 'GET'),
            ('/auth', 'GET'),
        ]
        
        # We should have at least this many routes identified
        assert len(expected_routes) >= 20
        
        # Check that we have a good mix of GET and POST
        get_routes = [r for r in expected_routes if r[1] == 'GET']
        post_routes = [r for r in expected_routes if r[1] == 'POST']
        assert len(get_routes) > 10
        assert len(post_routes) > 5
        
    def test_test_file_structure(self):
        """Test that our test files are properly structured."""
        import os
        
        # Check that test files exist
        test_files = [
            'tests/test_index_views.py',
            'tests/test_service_views.py', 
            'tests/test_credential_views.py',
            'tests/test_auth_and_misc.py',
            'tests/test_share_functionality.py'
        ]
        
        project_root = os.path.dirname(os.path.abspath(__file__))
        for test_file in test_files:
            full_path = os.path.join(project_root, '..', test_file)
            assert os.path.exists(full_path), f"Test file {test_file} should exist"
            
    def test_share_functionality_coverage(self):
        """Test that share functionality tests cover the key aspects."""
        # These are the key aspects our share tests should cover
        share_test_aspects = [
            'share_button_presence',
            'url_parameter_passing', 
            'javascript_function_inclusion',
            'special_character_handling',
            'different_modes',
            'error_handling'
        ]
        
        # We should test at least these aspects
        assert len(share_test_aspects) >= 5
